FIX: Do not allow 200 response with databaseConfig error in main.php
Previously, upon a fatal error in main.php, PHP would die and display an error
message to the user, but the HTTP response was still 200 OK. That has the
dangerous potential of poisoning caches with bad data. This change fixes the
behavior so that the page dies with a 500 Server Error code.