HHVM COMPATIBILITY: Send Proper HTTP Response Header
Silverstripe was sending a proper HTTP status header (i.e. 200 OK, 404 Not
Found), but HHVM was returning "explicit_header_response_code" to the browser.
This was being caused by subsequent calls to header() in HTTPResponse
overwriting the response code with just the integer portion of the code. HHVM,
seeing the lack of a code description, inputted it's default
"explicit_header_response_code". Removing the needless status overwriting fixes
this.