No need to empty DB when one was just created
Since a newly created table should be empty, there isn't really a need to empty
all the data from it. This save us from running an extra empty on the initial
database creation. It's not much but it is a start.

(NOTE: Cherry-picked in upgrade to 3.1.21)
Original commit: be6421cc185862a4e7f554d05bf8147cfebd3c59

Conflicts:
	dev/SapphireTest.php